requirements.txt
# SentinelVNC v2 Requirements
# Python 3.11 compatible packages

# Core ML & Data Science
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
shap==0.43.0
lime==0.2.0.1

# Deep Learning
tensorflow==2.15.0
keras==2.15.0

# Visualization & Dashboard
streamlit==1.28.1
Pillow==10.1.0
matplotlib==3.8.2
plotly==5.18.0

# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
pyotp==2.9.0
qrcode[pil]==7.4.2
python-dotenv==1.0.0

# Celery & Redis
celery==5.3.4
redis==5.0.1
flower==2.0.1

# Cryptography
cryptography==41.0.7

# Utilities
python-dateutil==2.8.2
pyyaml==6.0.1
tqdm==4.66.1
httpx==0.25.2

# ELK Stack Integration
elasticsearch==8.11.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.2

# Security Scanning
bandit==1.7.5
safety==2.3.5


